requirements.txt
gunicorn==19.9.0
boto3==1.9.220
Flask-MongoAlchemy==0.7.2
pymongo==3.9.0
dnspython==1.16.0
bcrypt==3.1.7
blinker==1.4
certifi==2019.6.16
cffi==1.12.3
Click==7.0
Flask==1.1.1
Flask-Bcrypt==0.7.1
Flask-Login==0.4.1
Flask-SQLAlchemy==2.4.0
Flask-WTF==0.14.2
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
Pillow==6.1.0
pycparser==2.19
pylint==2.3.1
six==1.12.0
SQLAlchemy==1.3.8
Werkzeug==0.15.5
WTForms==2.2.1
autopep8==1.4.4